requirements.txt
absl-py==0.1.13
astor==0.6.2
bleach==1.5.0
click==6.7
Flask==0.12.2
gast==0.2.0
grpcio==1.10.1
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
nltk==3.2.5
numpy==1.14.2
protobuf==3.5.2.post1
pycparser==2.18
regex==2018.2.21
six==1.11.0
tensorboard==1.7.0
tensorflow-gpu==1.7.0
termcolor==1.1.0
ujson==1.35
Werkzeug==0.14.1
